refactor: transition to modular whipsmart.nvim architecture with pack-manager
